+++ utils/python/utils_test.go
@@ -2,7 +2,7 @@ package utils
 
 import (
 	biutils "github.com/jfrog/build-info-go/utils"
-	"github.com/jfrog/jfrog-cli-core/v2/xray/audit"
+	"github.com/jfrog/jfrog-cli-core/v2/xray/commands/audit/sca"
 	"github.com/jfrog/jfrog-client-go/utils/io/fileutils"
 	"github.com/stretchr/testify/assert"
 	"path/filepath"
@@ -28,7 +28,7 @@ func initPoetryTest(t *testing.T) (string, func()) {
 	// Create and change directory to test workspace
 	testAbs, err := filepath.Abs(filepath.Join("..", "..", "xray", "commands", "testdata", "poetry-project"))
 	assert.NoError(t, err)
-	poetryProjectPath, cleanUp := audit.CreateTestWorkspace(t, "poetry-project")
+	poetryProjectPath, cleanUp := sca.CreateTestWorkspace(t, "poetry-project")
 	assert.NoError(t, biutils.CopyDir(testAbs, poetryProjectPath, true, nil))
 	return poetryProjectPath, cleanUp
 }
