--- xray/audit/jas/jasmanager_test.go
@@ -1,70 +1,0 @@
-package jas
-
-import (
-	"github.com/jfrog/jfrog-cli-core/v2/xray/utils"
-	"os"
-	"testing"
-
-	rtutils "github.com/jfrog/jfrog-cli-core/v2/artifactory/utils"
-	"github.com/jfrog/jfrog-cli-core/v2/utils/config"
-	"github.com/jfrog/jfrog-cli-core/v2/utils/coreutils"
-	"github.com/jfrog/jfrog-client-go/utils/io/fileutils"
-	"github.com/jfrog/jfrog-client-go/xray/services"
-	"github.com/stretchr/testify/assert"
-)
-
-var fakeBasicXrayResults = []services.ScanResponse{
-	{
-		ScanId: "scanId_1",
-		Vulnerabilities: []services.Vulnerability{
-			{IssueId: "issueId_1", Technology: coreutils.Pipenv.ToString(),
-				Cves:       []services.Cve{{Id: "testCve1"}, {Id: "testCve2"}, {Id: "testCve3"}},
-				Components: map[string]services.Component{"issueId_1_direct_dependency": {}, "issueId_3_direct_dependency": {}}},
-		},
-		Violations: []services.Violation{
-			{IssueId: "issueId_2", Technology: coreutils.Pipenv.ToString(),
-				Cves:       []services.Cve{{Id: "testCve4"}, {Id: "testCve5"}},
-				Components: map[string]services.Component{"issueId_2_direct_dependency": {}, "issueId_4_direct_dependency": {}}},
-		},
-	},
-}
-
-var mockDirectDependencies = []string{"issueId_2_direct_dependency", "issueId_1_direct_dependency"}
-var mockMultiRootDirectDependencies = []string{"issueId_2_direct_dependency", "issueId_1_direct_dependency", "issueId_3_direct_dependency", "issueId_4_direct_dependency"}
-
-var fakeServerDetails = config.ServerDetails{
-	Url:      "platformUrl",
-	Password: "password",
-	User:     "user",
-}
-
-func initJasTest(t *testing.T, workingDirs ...string) (*AdvancedSecurityScanner, func()) {
-	assert.NoError(t, rtutils.DownloadAnalyzerManagerIfNeeded())
-	scanner, err := NewAdvancedSecurityScanner(workingDirs, &fakeServerDetails)
-	assert.NoError(t, err)
-	return scanner, func() {
-		assert.NoError(t, scanner.scannerDirCleanupFunc())
-	}
-}
-
-func TestGetExtendedScanResults_AnalyzerManagerDoesntExist(t *testing.T) {
-	tmpDir, err := fileutils.CreateTempDir()
-	defer func() {
-		assert.NoError(t, fileutils.RemoveTempDir(tmpDir))
-	}()
-	assert.NoError(t, err)
-	assert.NoError(t, os.Setenv(coreutils.HomeDir, tmpDir))
-	defer func() {
-		assert.NoError(t, os.Unsetenv(coreutils.HomeDir))
-	}()
-	scanResults := &utils.ExtendedScanResults{XrayResults: fakeBasicXrayResults, ScannedTechnologies: []coreutils.Technology{coreutils.Yarn}}
-	err = RunScannersAndSetResults(scanResults, []string{"issueId_1_direct_dependency", "issueId_2_direct_dependency"}, &fakeServerDetails, nil, nil)
-	// Expect error:
-	assert.Error(t, err)
-}
-
-func TestGetExtendedScanResults_ServerNotValid(t *testing.T) {
-	scanResults := &utils.ExtendedScanResults{XrayResults: fakeBasicXrayResults, ScannedTechnologies: []coreutils.Technology{coreutils.Pip}}
-	err := RunScannersAndSetResults(scanResults, []string{"issueId_1_direct_dependency", "issueId_2_direct_dependency"}, nil, nil, nil)
-	assert.NoError(t, err)
-}
