+++ xray/commands/audit/jas/common_test.go
@@ -1,0 +1,23 @@
+package jas
+
+import (
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+func TestHideSecret(t *testing.T) {
+	tests := []struct {
+		secret         string
+		expectedOutput string
+	}{
+		{secret: "", expectedOutput: "***"},
+		{secret: "12", expectedOutput: "***"},
+		{secret: "123", expectedOutput: "***"},
+		{secret: "123456789", expectedOutput: "123************"},
+		{secret: "3478hfnkjhvd848446gghgfh", expectedOutput: "347************"},
+	}
+
+	for _, test := range tests {
+		assert.Equal(t, test.expectedOutput, hideSecret(test.secret))
+	}
+}
