+++ xray/commands/audit/sca/go/golang.go
